Changed layout of homepage
